renamed Myenums to UnitEnums and removed flush
