+++ components/driveTrain.py
@@ -1,4 +1,4 @@
-from utils.Myenum import positionUnits
+from utils.UnitEnums import positionUnits
 from enum import Enum, auto
 import math
 import wpilib.drive
