+++ robot.py
@@ -121,7 +121,6 @@ class MyRobot(MagicRobot):
         self.buttonManager.registerButtonEvent(self.xboxMap.drive, XboxController.Button.kBumperLeft, ButtonEvent.kOnRelease, self.driveTrain.disableCreeperMode)
         self.buttonManager.registerButtonEvent(self.xboxMap.drive, XboxController.Button.kBumperRight, ButtonEvent.kOnPress, self.navx.reset)
 
-        self.MXPserial.flush()
         self.driveTrain.resetDistTraveled()
 
         self.shooter.autonomousDisabled()
