+++ utils/motorHelper.py
@@ -3,7 +3,7 @@
 import rev
 import ctre
 import logging as log
-from .Myenum import positionUnits, velocityUnits
+from .UnitEnums import positionUnits, velocityUnits
 
 def createMotor(motorDescp, motors = {}):
     '''This is where all motors are set up.
